refactor: re-engineer Go CLI — idiomatic design, not a Python port
Previous implementation was a line-for-line translation: 9 source files,
7 test files, cobra for CLI, and a StorageClient interface that existed
purely to support mock-based tests.

This is a clean re-engineering from intent:

  - 3 source files (main.go, gcs.go, crypto.go) + 1 test file
  - stdlib flag.NewFlagSet replaces cobra (4 subcommands don't need a framework)
  - Single transfer() function shared by upload and pack — they are the
    same workflow; the only difference is file vs folder
  - No StorageClient interface, no mocks — GCS functions are direct calls;
    unit tests cover what is actually unit-testable (crypto, parsing,
    formatting); E2E is covered by test-run.sh
  - Security-critical dependencies retained: cloud.google.com/go/storage
    for V4 signed URL construction, github.com/yeka/zip for AES-256
    zip compatibility with 7-Zip

22 tests pass. Binary builds and runs.
